+++ genDSyms.jl
@@ -1,7 +1,7 @@
 include("dsyms.jl")
 
 
-for (count1, dset) in enumerate(DSetGenerator(2, parse(Int, ARGS[1])))
+for (count1, dset) in enumerate(DSetGenerator(2, parse(Int64, ARGS[1])))
     orbs = orbits(dset)
     vs = map(minV, orbs)
     curv = curvature(dset, orbs, vs)
